requirements.txt
discord
requests
python-dotenv
selenium
webdriver_manager